refactor: WebbProvider a bit for readability
