--- libs/api-provider-environment/src/constants.ts
@@ -1,5 +1,0 @@
-const APP_NAME = 'Webb Dapp';
-
-export default {
-  APP_NAME,
-};
