+++ libs/api-provider-environment/src/index.ts
@@ -1,7 +1,6 @@
 export * from './RouterProvider';
 export * from './WebbProvider';
 export * from './app-event';
-export * from './constants';
 export * from './error';
 export * from './modal-queue-manager';
 export * from './store';
